pyproject.toml
[tool.poetry]
name = "homematch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lancedb = "^0.6.1"
transformers = "^4.38.2"
scrapy = "^2.11.1"
ipykernel = "^6.29.3"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
